size_t -> int

+++ dali/core/geom_mat_test.cu
@@ -21,8 +21,8 @@ namespace dali {
 
 TEST(Mat, Constructor) {
   mat3x4 m = {{ { 1, 2, 3, 4 }, { 5, 6, 7, 8 }, { 9, 10, 11, 12 } }};
-  for (size_t i = 0; i < m.rows; i++)
-    for (size_t j = 0; j < m.cols; j++)
+  for (int i = 0; i < m.rows; i++)
+    for (int j = 0; j < m.cols; j++)
       EXPECT_EQ(m(i, j), i*m.cols+j+1);
 
   mat<3, 1> m31 = vec3(1, 2, 3);
@@ -36,8 +36,8 @@ TEST(Mat, Equality) {
   imat2x4 b = {{ { 1, 2, 3, 4 }, { 5, 6, 7, 8 } }};
   EXPECT_TRUE((a == b));
   EXPECT_FALSE((a != b));
-  for (size_t i = 0; i < a.rows; i++) {
-    for (size_t j = 0; j < a.cols; j++) {
+  for (int i = 0; i < a.rows; i++) {
+    for (int j = 0; j < a.cols; j++) {
       a(i, j)+=10;
       EXPECT_FALSE((a == b));
       EXPECT_TRUE((a != b));
@@ -55,8 +55,8 @@ TEST(Mat, FromScalar) {
     { 0, 5, 0, 0 },
     { 0, 0, 5, 0 },
   }};
-  for (size_t i = 0; i < m.rows; i++)
-    for (size_t j = 0; j < m.cols; j++)
+  for (int i = 0; i < m.rows; i++)
+    for (int j = 0; j < m.cols; j++)
       EXPECT_EQ(m(i, j), ref(i, j));
 }
 
@@ -67,8 +67,8 @@ TEST(Mat, Transpose) {
     { 9, 10, 11, 12 },
   }};
   mat4x3 transposed = m.T();
-  for (size_t i = 0; i < m.rows; i++)
-    for (size_t j = 0; j < m.cols; j++)
+  for (int i = 0; i < m.rows; i++)
+    for (int j = 0; j < m.cols; j++)
       EXPECT_EQ(transposed(j, i), m(i, j));
 }
 
@@ -90,8 +90,8 @@ TEST(Mat, Mul) {
     { 12+200+2800+36000, 4+80+1200+16000 }
   }};
   mat3x2 result = m1*m2;
-  for (size_t i = 0; i < 3; i++)
-    for (size_t j = 0; j < 2; j++)
+  for (int i = 0; i < 3; i++)
+    for (int j = 0; j < 2; j++)
       EXPECT_EQ(result(i, j), ref(i, j)) << "@ " << i << ", " << j;
 }
 
@@ -116,8 +116,8 @@ TEST(Mat, Mul4x4) {
     { 55804, 60248, 64692, 69136 }
   }};
   mat4 result = m1 * m2;
-  for (size_t i = 0; i < 4; i++)
-    for (size_t j = 0; j < 4; j++)
+  for (int i = 0; i < 4; i++)
+    for (int j = 0; j < 4; j++)
       EXPECT_EQ(result(i, j), ref(i, j)) << "@ " << i << ", " << j;
 }
 
@@ -126,8 +126,8 @@ TEST(Mat, CompoundAssignMatrix) {
   #define TEST_ASSIGN_OP(op)                          \
   {                                                   \
     m1 op##= m2;                                      \
-    for (size_t i = 0; i < 3; i++)                    \
-      for (size_t j = 0; j < 4; j++)                  \
+    for (int i = 0; i < 3; i++)                    \
+      for (int j = 0; j < 4; j++)                  \
         EXPECT_EQ(m1(i, j), orig(i, j) op m2(i, j));  \
     m1 = orig;                                        \
   }
@@ -188,8 +188,8 @@ TEST(Mat, CompoundAssignScalar) {
 #define TEST_ASSIGN_OP(op)                    \
   {                                           \
     m1 op## = s;                              \
-    for (size_t i = 0; i < 3; i++)            \
-      for (size_t j = 0; j < 4; j++)          \
+    for (int i = 0; i < 3; i++)            \
+      for (int j = 0; j < 4; j++)          \
         EXPECT_EQ(m1(i, j), orig(i, j) op s); \
     m1 = orig;                                \
   }
@@ -235,11 +235,11 @@ struct cuda_rng {
   unsigned seed;
 };
 
-template <size_t rows, size_t cols, typename T, typename RNG>
+template <int rows, int cols, typename T, typename RNG>
 DALI_HOST_DEV
 void RandomFill(mat<rows, cols, T> &m, RNG &rng, int lo, int hi) {
-  for (size_t i = 0; i < rows; i++)
-    for (size_t j = 0; j < cols; j++)
+  for (int i = 0; i < rows; i++)
+    for (int j = 0; j < cols; j++)
       m(i, j) = rng()%(hi-lo) + lo;
 }
 
@@ -250,17 +250,17 @@ TEST(Mat, BinOp) {
     cuda_rng rng = { 12345 };                                 \
     RandomFill(m1, rng, lo, hi); RandomFill(m2, rng, lo, hi); \
     mat_type result = m1 op m2;                               \
-    for (size_t i = 0; i < result.rows; i++)                  \
-      for (size_t j = 0; j < result.cols; j++)                \
+    for (int i = 0; i < result.rows; i++)                  \
+      for (int j = 0; j < result.cols; j++)                \
         EXPECT_EQ(result(i, j), m1(i, j) op m2(i, j));        \
     int scalar = rng()&31;                                    \
     result = m1 op scalar;                                    \
-    for (size_t i = 0; i < result.rows; i++)                  \
-      for (size_t j = 0; j < result.cols; j++)                \
+    for (int i = 0; i < result.rows; i++)                  \
+      for (int j = 0; j < result.cols; j++)                \
         EXPECT_EQ(result(i, j), m1(i, j) op scalar);          \
     result = scalar op m2;                                    \
-    for (size_t i = 0; i < result.rows; i++)                  \
-      for (size_t j = 0; j < result.cols; j++)                \
+    for (int i = 0; i < result.rows; i++)                  \
+      for (int j = 0; j < result.cols; j++)                \
         EXPECT_EQ(result(i, j), scalar op m2(i, j));          \
   }
 
@@ -341,8 +341,8 @@ TEST(Mat, CatCols) {
     { 4, 5, 6, 30, 40 },
     { 7, 8, 9, 50, 60 }
   }};
-  for (size_t i = 0; i < 3; i++)
-    for (size_t j = 0; j < 5; j++)
+  for (int i = 0; i < 3; i++)
+    for (int j = 0; j < 5; j++)
       EXPECT_EQ(result(i, j), ref(i, j)) << "@ " << i << ", " << j;
 }
 
@@ -360,8 +360,8 @@ TEST(Mat, CatColsVec) {
     { 4, 5, 6, 20 },
     { 7, 8, 9, 30 }
   }};
-  for (size_t i = 0; i < 3; i++)
-    for (size_t j = 0; j < 4; j++)
+  for (int i = 0; i < 3; i++)
+    for (int j = 0; j < 4; j++)
       EXPECT_EQ(result(i, j), ref(i, j)) << "@ " << i << ", " << j;
 
   result = cat_cols(b, a);
@@ -370,8 +370,8 @@ TEST(Mat, CatColsVec) {
     { 20, 4, 5, 6 },
     { 30, 7, 8, 9 }
   }};
-  for (size_t i = 0; i < 3; i++)
-    for (size_t j = 0; j < 4; j++)
+  for (int i = 0; i < 3; i++)
+    for (int j = 0; j < 4; j++)
       EXPECT_EQ(result(i, j), ref(i, j)) << "@ " << i << ", " << j;
 
   mat3x2 result2 = cat_cols(b, c);
@@ -380,8 +380,8 @@ TEST(Mat, CatColsVec) {
     { 20, 50 },
     { 30, 60 }
   }};
-  for (size_t i = 0; i < 3; i++)
-    for (size_t j = 0; j < 2; j++)
+  for (int i = 0; i < 3; i++)
+    for (int j = 0; j < 2; j++)
       EXPECT_EQ(result2(i, j), ref2(i, j)) << "@ " << i << ", " << j;
 }
 
@@ -403,8 +403,8 @@ TEST(Mat, CatRows) {
     { 10, 20, 30 },
     { 40, 50, 60 }
   }};
-  for (size_t i = 0; i < 5; i++)
-    for (size_t j = 0; j < 3; j++)
+  for (int i = 0; i < 5; i++)
+    for (int j = 0; j < 3; j++)
       EXPECT_EQ(result(i, j), ref(i, j)) << "@ " << i << ", " << j;
 }
